Internals: Further performance improvement of AstNode type tests
In this patch we are replacing the virtual type() method on AstNode
with a non-virtual, inlined accessor to a const member variable m_type.
This means that in order to be able to use this for type testing, it
needs to be initialized based on the final type of the node. This is
achieved by passing the relevant AstType value back through the
constructor call chain. Most of the boilerplate involved is auto
generated by first feeding V3AstNodes.h through astgen to get
V3AstNodes__gen.h, which is then included in V3Ast.h. No client code
needs to be aware and there is no functional change intended.

Eliminating the virtual function call to fetch the node type identifier
results in measured compilation speed improvement of 5-10% as it
eliminates up to 20% of all mispredicted branches from the execution.