Add FreeBSD to Travis CI.
This adds FreeBSD to the Travis CI config. As ccache doesn't work for
FreeBSD on Travis at the moment, this just builds and runs `make
examples`, rather than running the full tests.

The CI shell scripts are updated with a portable shebang and to use the
$MAKE environment variable. The `ci/test.sh` script is updated for
consistency, even though it is not run on FreeBSD as part of this PR.

The FreeBSD job is added as `!= cron` in this commit, so we can see it
run as part of the PR's CI. Once it passes and we're happy, I can add a
subsequent commit to make it `= cron`.