Add .dot graph visualization of AstExecGraph
This commit adds the creation of .dot files - used by GraphViz - to visualize how mtasks are statically scheduled across the set of specified threads.
We visualize each thread as a row, with nodes of a row being the mtasks scheduled for the given thread. The width of the mtask nodes are proportional to their cost. mtask dependencies are shown using an edge between the source and sink mtasks.

It could be argued that this information would be more easily displayed in a .vcd file. However, i find this rather restrictive in terms of not being able to annotate ie. mtask dependencies. Furthermore, keeping this to .dot files allows for embedding more details in the future, if the multithread scheduling implementation is elaborated.