Recurse over task dependencies when writing AstExecGraph .dot file
It may happen that an mtask's in edges are not ordered as they appear when iterating
over the m_depGraphp. As such, previously, a situation could occur where a dependent
node was not present in the mtaskRhsEdge map, resulting in an incorrect
visualization.
Now, we recursively log all dependencies of an mtask before logging the mtask itself.