Add computation of critical path to AstExecGraph
To compute a critical path estimate, we first locate the mtask which is expected to finish the latest. From here, we backtrack through its dependencies, selecting the mtask which finish the _latest_ in time.
This is done instead of checking which dependencies finish exactly when a critical task starts, due to the fact that scheduling may choose to schedule a task some time after a dependency finished executing on another thread (see PartPackMTasks).

Required for #2779