Fix slice-assign overflow bug (Issue 2803)
When assigning to a bit-select (array slice), there is nothing stopping
the bit-select referring to bits beyond the width of the signal we're
actually assigning to.  Because the expansion of this logic (V3Expand
and verilated.h) happens after V3Clean, this bypasses the normal
cleaning process and can result in a signal containing a value whose
width exceeds the signal width.

A lot of codepaths cover the same logic depending on the widths
involved, which things are constant, and whether the logic is done by
verilated.h or expanded by V3Expand.  This commit adds regression tests
for this bug in each codepath and fixes the bug in all cases.

This fixes #2803