Fix __DOT__ concat of hashed names
In VName::hashedName() the names of some AST nodes are hashed if they
are too long, this takes the form of truncating them then appending a
suffix of the form __Vhsh<hash>.

Later, for the purposes of tracing they are dehashed. Dehashing involves
looking up the suffix in a map which was populated when the original
name was hashed. An assert occurs if the suffix is not found in the map.
The indication of whether this dehashing is necessary is whether the
string contains __Vhsh.

Several places in the code name new nodes by concatenating the literal
__DOT__ to an existing node name plus a new identifier. However, if the
existing node name has been hashed this results in a new node name of
the form long_string__Vhsh<hash>__DOT__ident. When generating code to
support tracing these new names appear as though they require dehashing
as they contain __Vhsh, but they do not appear in the map as they
weren't generated by the hashedName function, triggering the assert.

The attempted fix is to, at dehash, split the name into components
separated by __DOT__, dehash those in isolation, then reconcatenate with
__DOT__.