Fix evaluation when variables are assigned in DPI exported functions
Logic downstream of variables assigned in a DPI exported function (that
may be called from outside eval) were not re-evaluated properly (see
test added in this patch) as we didn't realize the variables might be
volatile. This is fixed by the following changes in V3Order:

1. Walk all DPI exported functions, and add an edge to the ordering
graph from the input vertex, to all variables written by the exported
function (except for function locals).

2. Assign all variable vertices that have an in-edge from the input
vertex to the combinational domain.

Note that point 2 above does not cause any change for variables that
were already descendants of the input node, as they would have already
been assigned to the combinational domain due to being inputs (or
nonOutputs) of the top level module.