Escape identifiers that contain invalid characters
When translating VHDL to Verilog, GHDL may produce identifiers with ':'
in them. As required by the Verilog spec, it escapes the identifer with
a leading backslash.

When verilator produces a VCD file based on this Verilog, the leading
backslash is missing. Gtkwave fails to parse such a file.

Escape any identifier that is not a combination of letters, digits, '$'
or '_'.