fst: include vector range to allow filling with 0's in gtkwave
Verilator currently names FST signals by the node name and if the signal
is an array, appending the index number in square brackets. This works
fine, but unlike VCD, it means that a signal defined as say:

  logic [31:2] addr;

ends up in the FST as a [29:0] instead and it is not possible to use
GTKWave's range fill to pad the LSB's with 0's making it necessary to
calculate the byte address.  When we have a bussed signal, include the
full range in the signal name so that analysis tools can pad the LSBs.

Fixes #3296

Signed-off-by: Jamie Iles <jamie@jamieiles.com>