Lazy module body parametrization
This is the first patch of a series designed to implement parametrized
classes support, and might as well help to support nested modules in the
future.

It makes sure that the whole body of the module/class is processed
(grabbing all the references to parametrized modules) before in-module
classes (or possibly---modules) have their bodies deparametrized.
This avoids wrongly applying default values of parameters to any
expressions that need to be constified during the V3Param phase.

The influenced test outputs are included for bisecting reliability,
they will be removed in following commits.