Internals: less strict nodeDeparam for later reuse
This commit makes cellDeparam accept any AstNode that fits here.  It
enables accepting ClassRefDType and ClassOrPackageRef in the following
commits.  Another way to achieve this would be to introduce a new C++
type (interface) that would be implemented by all of them.  This is a
bit hard, as ClassRefDType must inherit NodeDType and the others must
not, and other NodeDTypes should not implement the new interface.