Allow module lists to contain other statements
If a class (or module) is an element of a module body, the linked list
containing it may contain other types of statements as well.  This
commit relaxes the assumption and makes sure the parametrized module to
be inserted on a linked list just stops on the first non-module
statement when searching for modules with the right level.