Consider nested deparametrization
This can happen for example when parsing expressions such as:

    Cls#(Cls#(8)::PBASE) x;

Iterating over cell pins can lead to pushing a ClassOrPackageRef onto
the queue, but I was unable to reproduce any actual segfaults.
The support for such expressions is out of reach currently, as it would
require factoring out parts of V3LinkDot.

Such cases currently fail with a unique internal error of
V3Const::constify not supporting AstDot expressions.

Co-authored-by: Krzysztof Bieganski <kbieganski@antmicro.com>