Force printing fields from superclasses
Ordering of classes is mostly unreliable, so just stringify the subclass
even if it is not known to have fields.  This can be reverted some day
when class definitions are sorted properly.

While this might add runtime testbench performance overhead, the feature
never worked correctly, so we just stick to it for now.