Rework to handle strings of any size
After this, the saved old value is interpretted as a std::string, and a list of where those are is used to init/destroy them. This removes any small-string opts.

Further, the max string size is used. If the size is exceeded, the first "size" bytes are still put into the trace. The whole value is stored for deduping purposes (ie. strings that differ past the given size will show as a change, even if the changed bytes aren't in the trace).

Compiles, needs testing.