Enable access to constructs inside type parameters
This changeset is supposed to bring support for accesses like:

	class Cls#(type TYPE1);
	   ...
	    TYPE1::some_method();
	   ...
	endclass

It is done by delaying dot resolution on type parameters until they get
resolved by V3Param, and doing a more thorough reference skip.

There is an extra change to debug messages that could be separated from
here, but it has helped a lot during the debugging, so it should
probably be kept.