Added test for intefaces with incorrect localparam values
It appears that the value of symbolsPerBeatDivBy2 is not recomputed for new interfaces.

```
top.t.intf:              symbolsPerBeat  16, symbolsPerBeatDivBy2  8, mismatch 0
top.t.theCore.core_intf: symbolsPerBeat  64, symbolsPerBeatDivBy2  8, mismatch 1
```

In this case 'intf' is created and symbolsPerBeatDivBy2 is extracted from the interface
in a submodule. Once this is done future interfaces of the same type (i.e. 'core_intf' above)
seem to inherit the symbolsPerBeatDivBy2 computed for the first declared interface ('intf')
even though the symbolsPerBeat (which determines the correct value for symbolsPerBeatDivBy2)
value is different for the second instantiated interface.