V3EmitMk.cpp: Makefile generator escape for CFLAGS/LDFLAGS/Mdir
When generating *.mk try to preserve the original verilator command
line / input argument to present to the toolchain verbatim.

This patch takes the view that any interpolation of the input
arguments by verilator, make or the shell executing the toolchain
command should be reduced to near zero or at least become a
documented feature / caveat.

It is speculated the most common use for this change will be the
SPACE character escaping that is common in Windows platform paths
that can be present in the paths of C++ SDKs installed on that
platform.

Also to some degree the Windows platform backslash path delimiters,
both make and the shell are lenient in their interpretation of
unnecessary escape characters which allow users to be unaware of
corner cases.  So after this patch there will be a visible
difference of emitted values in the form dir\\file but doing this
closes many corner cases.

It is possible some users may be affected by this change, because
they have already put in place workarounds to transform their
verilator input arguments to gain a desired outcome when the
arguments are presented to the toolchain.  Use of such workarounds
are more brittle and a less portable (relying on a specific build
setup on a specific platform) than exercising the general rule:
 verilator should try to preserve arguments

Resolving such cases should be a matter of removing any additional
escaping you are performing on the input arguments.

It would be considered a bug of this patch if you are unable to
achieve the original toolchain argument presentation outcome by
any means of modifying the input arguments.
It would be considered a bug of this patch, if you are able to
observe additional escape characters being presented to the
toolchain due to this patch.