bin/verilator $VERILATOR_BIN handling improvement
Feature #1:

If $VERILATOR_BIN is already an absolute path use the path as-is such as:
  /usr/bin/verilator_bin
  /usr/local/bin/verilator_bin
  /opt/verilator/bin/verilator_bin
  /opt/verilator_v5.010/bin/verilator_bin

Feature #2:

If 'make install' has been done, and a default VERILATOR_ROOT path is hardcoded.
Allow the original installation to be relocated/moved to another path on the system.
When $VERILATOR_ROOT is then setup and bin/verilator is in $PATH we expect
it to just work out of the box.

The patch still allows a binary in the $VERILATOR_ROOT/bin/verilator_bin
to have priority.  But when it is not found there take a look for
$VERILATOR_ROOT/../../bin/verilator_bin

This path is the equivalent to searching at:
   /usr/local/share/verilator/../../bin/verilator_bin

This is the equivalent location of a default 'make install' into /usr/local.

Then moving all the installed files into /opt/verilator_v5.010 creating paths such as:
   /opt/verilator_v5.010/share/verilator/include/verilated.h
   /opt/verilator_v5.010/bin/verilator_bin

This is really useful for Docker like environmments and system packaging scripts
to be able to build and install using standard settings.
Extract all files written during the install into an image resulting docker image,
this resulting image can then be mounted at any location of any other environment
and expect to work out of the box providing you have setup $VERILATOR_ROOT and
you have the verilator bootstrap Perl script in the $PATH.