Make node iteration stricter
This is a prep patch for an experiment on removing m_iterpp (which I
have in a hacked up form, and does not actually help much, 1.2% memory
saving, yay...). Even if we don't land that, I think we want these
changes that makes iteration a little bit stricter. Notably, there are
assertions (only under VL_DEBUG, due to being in hot code), that ensure
that we don't recursively iterate the same node. We used to do this in a
few places (which I fixed), but this is generally not safe due to the
recursive call to iterateAndNext overwriting m_iterpp, and hence any
subsequent edits would be unsafe.

Apart from the direct recursion, we also need to prevent
addHereThisAsNext from updating the iteration pointer. I think this is a
good thing. Previously it used to force iteration to restart on the
nodes inserted *before* the currently iterated node, which would also
mean that the currently iterated node would then be visited again, as it
became a successor of the inserted node. With this patch, iteration
continues with the edited node only if it's replacing the current node,
or if the new node is a successor (nextp) of the current node, but not
if the edit is adding the node before the current node. I fixed the few
places (V3Premit, V3Task) where we used to rely on the old behaviour.
The fix is simply to explicitly iterate the new nodes you insert before
the current node.

Patch is performance neutral.