Fix unsafe write in _vl_insert_WI.
As an example: if the destination is [11:0][7:0] = 96 bits,
which is an even multiple of the 32-bit chunk size that Verilator
uses to break up signals wider than 64-bits, then if you try to
write bits [111:80], then 2 bytes go into [95:80] (lword) and two
bytes would, prior to this commit, try to write into [111:96] (hword).

This write to hword is an unsafe/illegal write since the signal is
split up as [2:0][31:0], so hword = 3 would effectively clobber whatever
follows your 3-element uint32 array in the C++ struct generated by
verilator; in the example testbench it clobbers d_bytes, making it
zero.

https://github.com/verilator/verilator/issues/4850